requirements.txt
# Akashic Records - AI Code Index System
# Python 3.10+

# Vector DB
qdrant-client>=1.7.0

# Embedding (via llama.cpp server - HTTP client only)
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# Reranker
sentence-transformers>=2.2.0
torch>=2.0.0

# Code Parsing
tree-sitter>=0.21.0
tree-sitter-c-sharp>=0.21.0
tree-sitter-cpp>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0

# MCP Server
mcp>=1.0.0
starlette>=0.36.0
uvicorn>=0.27.0

# Metadata DB
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
watchdog>=3.0.0
click>=8.1.0
rich>=13.0.0
psutil>=5.9.0

# Async support
anyio>=4.0.0
